little_grid_world/pyproject.toml
[tool.poetry]
name = "little_grid_world"
version = "0.0.1"
description = ""
authors = ["william.murray <william.murray@automotivepower.com>"]

[tool.poetry.dependencies]
python = "3.11.1"
Flask = "3.0.0"
Flask-SocketIO = "5.3.6"
boto3 = "1.28.73"
Pillow = "10.1.0"
eventlet = "^0.33.3"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
vulture = "2.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "atp_v3.atp_python_v1:main"